Fix 64-bit-related warnins.

+++ mux/src/htab.cpp
@@ -1,6 +1,6 @@
 // htab.cpp -- Table hashing routines.
 //
-// $Id: htab.cpp,v 1.13 2004-07-07 18:00:08 sdennis Exp $
+// $Id: htab.cpp,v 1.14 2004-07-09 15:16:00 sdennis Exp $
 //
 // MUX 2.4
 // Copyright (C) 1998 through 2004 Solid Vertical Domains, Ltd. All
@@ -67,7 +67,7 @@ void *hashfindLEN(const void *str, size_t nStr, CHashTable *htab)
  * * hashaddLEN: Add a new entry to a hash table.
  */
 
-int hashaddLEN(const void *str, int nStr, void *hashdata, CHashTable *htab)
+int hashaddLEN(const void *str, size_t nStr, void *hashdata, CHashTable *htab)
 {
     // Make sure that the entry isn't already in the hash table.  If it
     // is, exit with an error.
@@ -85,7 +85,7 @@ int hashaddLEN(const void *str, int nStr, void *hashdata, CHashTable *htab)
     {
         HP_HEAPLENGTH nRecord;
         htab->Copy(iDir, &nRecord, &htab_rec);
-        int nTarget = nRecord - sizeof(int *);
+        size_t nTarget = nRecord - sizeof(int *);
 
         if (  nTarget == nStr
            && memcmp(str, htab_rec.aTarget, nStr) == 0)
